Changed ClassHours to type double

+++ PlannabelleClassLibrary/Models/Module.cs
@@ -12,6 +12,6 @@ namespace PlannabelleClassLibrary.Models
         public string Code { get; set; } = null!;
         public string Name { get; set; } = null!;
         public int Credits { get; set; }
-        public int ClassHoursPerWeek { get; set; }
+        public double ClassHoursPerWeek { get; set; }
     }
 }
